Add the conversion
